:recycle: Merge the 2 examples
